// src/app/demo/components/apps/job-tracker/models/job-application.ts
export interface JobApplication {
    pk?: number;  // Optional because it's auto-generated by the database
    company_name: string;
    job_title: string;
    job_location?: string;
    job_posting_link?: string;
    job_source?: string;
    date_saved?: Date;
    date_applied?: Date;
    application_status?: string;
    follow_up_date?: Date;
    contact_person?: string;
    contact_email?: string;
    application_notes?: string;
    interview_rounds?: number;
    interview_dates?: Date[];
    interview_feedback?: string;
    offer_received?: boolean;
    salary_offered?: number;
    additional_perks?: string;
    job_accepted?: boolean;
    reason_for_declining?: string;
    created_at?: Date;
}